+++ src/main/java/io/github/jpdev/asaassdk/doc/Examples.java
@@ -17,6 +17,7 @@ import io.github.jpdev.asaassdk.rest.myaccount.fee.AccountFee;
 import io.github.jpdev.asaassdk.rest.myaccount.status.MyAccountStatus;
 import io.github.jpdev.asaassdk.rest.notification.NotificationConfig;
 import io.github.jpdev.asaassdk.rest.payment.Payment;
+import io.github.jpdev.asaassdk.rest.payment.PaymentCreator;
 import io.github.jpdev.asaassdk.rest.payment.children.SplitSetting;
 import io.github.jpdev.asaassdk.rest.payment.enums.PaymentLinkChargeType;
 import io.github.jpdev.asaassdk.rest.payment.enums.PaymentStatus;
@@ -33,91 +34,39 @@ import io.github.jpdev.asaassdk.rest.pix.qrcode.PixQrCode;
 import io.github.jpdev.asaassdk.rest.pix.qrcode.decode.PixDecodedQrCode;
 import io.github.jpdev.asaassdk.rest.pix.transaction.PixTransaction;
 import io.github.jpdev.asaassdk.rest.pix.transaction.PixTransactionReader;
-import io.github.jpdev.asaassdk.rest.pixautomatic.authorization.PixAutomaticAuthorization;
-import io.github.jpdev.asaassdk.rest.pixautomatic.authorization.immediate.ImmediateQrCodeCreator;
-import io.github.jpdev.asaassdk.rest.pixautomatic.paymentinstruction.PixAutomaticPaymentInstruction;
 import io.github.jpdev.asaassdk.rest.subscription.*;
 import io.github.jpdev.asaassdk.rest.transfer.Transfer;
 import io.github.jpdev.asaassdk.rest.transfer.children.*;
 import io.github.jpdev.asaassdk.utils.BillingType;
 import io.github.jpdev.asaassdk.utils.Money;
-import io.github.jpdev.asaassdk.utils.PixAutomaticAuthorizationFrequency;
 
 import java.math.BigDecimal;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.List;
-import java.util.Random;
+import java.util.*;
 
 public class Examples {
 
     public static void main(String[] args) {
         Asaas.initSandbox(Secret.getAccessToken()); // Initialize the SDK with your access token
-        createPixAuthorization();
-        readPixAutomaticAuthorizations();
-        readPaymentInstructions();
-    }
-
-    private static void createPixAuthorization() {
-        Calendar calendar = Calendar.getInstance();
-        calendar.add(Calendar.DAY_OF_MONTH, 1);
-        Date startDate = calendar.getTime();
-
-        ImmediateQrCodeCreator immediateQrCodeCreator = new ImmediateQrCodeCreator()
-            .setDescription("teste")
-            .setExpirationSeconds(3600)
-            .setOriginalValue(Money.create(10));
-
-        String contractId = String.valueOf(System.currentTimeMillis());
-        PixAutomaticAuthorization authorization = PixAutomaticAuthorization.creator()
-            .setCustomerId("cus_000007258649")
-            .setStartDate(startDate)
-            .setFrequency(PixAutomaticAuthorizationFrequency.MONTHLY)
-            .setValue(Money.create(100))
-            .setContractId("CONTRACT_IIDD" + contractId)
-            .setImmediateQrCode(immediateQrCodeCreator)
-            .create();
-
-        assert authorization.getId() != null;
-        assert authorization.getImmediateQrCode().getExpirationDate() != null;
-    }
-
-    private static void readPixAutomaticAuthorizations() {
-        for (PixAutomaticAuthorization authorization : PixAutomaticAuthorization.reader().read().getData()) {
-            assert authorization.getContractId() != null;
-        }
-    }
-
-    private static void readPaymentInstructions() {
-        List<PixAutomaticPaymentInstruction> paymentInstructionList = PixAutomaticPaymentInstruction.reader().read().getData();
-        paymentInstructionList.forEach(it -> {
-            assert it.getId() != null;
-            assert it.getDueDate() != null;
-        });
-
-        PixAutomaticPaymentInstruction fetchedPaymentInstruction = PixAutomaticPaymentInstruction.fetcher(paymentInstructionList.get(0).getId()).fetch();
-        assert fetchedPaymentInstruction.getPaymentId() != null;
-        assert fetchedPaymentInstruction.getDueDate() != null;
-        assert fetchedPaymentInstruction.getAuthorization() != null;
+        PixAutomaticExamples.start();
     }
 
     private static void createPaymentWithSplit() {
         Account account = createFirstAccountIfNecessary();
 
         SplitSetting split1 = new SplitSetting()
-                .setWalletId(account.getWalletId())
-                .setFixedValue(Money.create(10));
+            .setWalletId(account.getWalletId())
+            .setFixedValue(Money.create(10));
 
         CustomerAccount customerAccount = CustomerAccount.reader().read().getData().get(0);
 
         Payment payment = Payment.creator()
-                .setCustomer(customerAccount.id)
-                .setBillingType(BillingType.PIX)
-                .setDueDate(new Date())
-                .addSplit(split1)
-                .setValue(Money.create(100))
-                .setDescription("Teste")
-                .create();
+            .setCustomer(customerAccount.id)
+            .setBillingType(BillingType.PIX)
+            .setDueDate(new Date())
+            .addSplit(split1)
+            .setValue(Money.create(100))
+            .setDescription("Teste")
+            .create();
 
         System.out.println(payment.getSplit().get(0).getId().toString());
 
@@ -136,113 +85,113 @@ public class Examples {
         PixTransaction pixTransaction = PixTransaction.fetcher("bc515f74-d5c7-4bc2-93e5-3bafc0a9b15d").fetch();
         PixTransaction cancelledPixTransaction = PixTransaction.canceller("35363f6e-93e2-11ec-b9d9-96f4053b1bd4").create();
         ResourceSet<PixTransaction> pixTransactionDebitResourceSet = PixTransaction.reader()
-                .setType(PixTransactionType.DEBIT)
-                .read();
+            .setType(PixTransactionType.DEBIT)
+            .read();
     }
 
     private static void pixAddressKey() {
         ResourceSet<PixAddressKey> pixAddressKeyResourceSet = PixAddressKey.reader()
-                .setStatus(PixAddressKeyStatus.ACTIVE)
-                .setLimit(1)
-                .read();
+            .setStatus(PixAddressKeyStatus.ACTIVE)
+            .setLimit(1)
+            .read();
         PixAddressKey.creator().setType(PixAddressKeyType.EVP).create();
         PixAddressKey.reader().read();
     }
 
     private static void decodePixQrCode() {
         PixDecodedQrCode decodedQrCode = PixDecodedQrCode.decoder()
-                .setPayload("payload")
-                .create();
+            .setPayload("payload")
+            .create();
     }
 
     private static void transfer() {
         Transfer transfer = Transfer.pixAddressKeyCreator()
-                .setPixAddressKey("+5547999999999")
-                .setValue(Money.create(0.01))
-                .setDescription("teste")
-                .setPixAddressKeyType(PixAddressKeyType.PHONE)
-                .create();
+            .setPixAddressKey("+5547999999999")
+            .setValue(Money.create(0.01))
+            .setDescription("teste")
+            .setPixAddressKeyType(PixAddressKeyType.PHONE)
+            .create();
         System.out.println(transfer.getValue().toString());
 
         ResourceSet<Transfer> transferList = Transfer.reader().read();
 
         Date birthDate = new Date();
         BankAccountSetting bankAccountSetting = new BankAccountSetting()
-                .setBank(
-                        new BankSetting().setCode("085")
-                )
-                .setAccountName("Paulo")
-                .setOwnerName("Paulo")
-                .setOwnerBirthDate(new Date())
-                .setCpfCnpj("06928316000124")
-                .setAgency("0108")
-                .setAccount("10895")
-                .setAccountDigit("5")
-                .setBankAccountType(BankAccountType.CONTA_CORRENTE);
+            .setBank(
+                new BankSetting().setCode("085")
+            )
+            .setAccountName("Paulo")
+            .setOwnerName("Paulo")
+            .setOwnerBirthDate(new Date())
+            .setCpfCnpj("06928316000124")
+            .setAgency("0108")
+            .setAccount("10895")
+            .setAccountDigit("5")
+            .setBankAccountType(BankAccountType.CONTA_CORRENTE);
         Transfer pixManualTransfer = Transfer.pixManualCreator()
-                .setBankAccount(bankAccountSetting)
-                .setValue(Money.create(new BigDecimal(1.01)))
-                .create();
+            .setBankAccount(bankAccountSetting)
+            .setValue(Money.create(new BigDecimal(1.01)))
+            .create();
 
         System.out.println(pixManualTransfer.getValue().toString());
 
         Transfer ted = Transfer.tedCreator()
-                .setBankAccount(bankAccountSetting)
-                .setValue(Money.create(new BigDecimal(1.01)))
-                .create();
+            .setBankAccount(bankAccountSetting)
+            .setValue(Money.create(new BigDecimal(1.01)))
+            .create();
 
         Transfer internal = Transfer.internalCreator()
-                .setValue(Money.create(new BigDecimal(10)))
-                .setWalletId("0021c712-d963-4d86-a59d-031e7ac51a2e")
-                .create();
+            .setValue(Money.create(new BigDecimal(10)))
+            .setWalletId("0021c712-d963-4d86-a59d-031e7ac51a2e")
+            .create();
     }
 
     private static void recurringTransfer() {
         PixRecurring recurring = new PixRecurring()
-                .setFrequency(PixRecurringFrequency.MONTHLY)
-                .setQuantity(2);
+            .setFrequency(PixRecurringFrequency.MONTHLY)
+            .setQuantity(2);
 
         Transfer transfer = Transfer.pixAddressKeyCreator()
-                .setPixAddressKey("+5547999999999")
-                .setValue(Money.create(0.01))
-                .setDescription("teste")
-                .setPixAddressKeyType(PixAddressKeyType.PHONE)
-                .setRecurring(recurring)
-                .create();
+            .setPixAddressKey("+5547999999999")
+            .setValue(Money.create(0.01))
+            .setDescription("teste")
+            .setPixAddressKeyType(PixAddressKeyType.PHONE)
+            .setRecurring(recurring)
+            .create();
         System.out.println(transfer.getRecurring());
     }
 
     private static void bill() {
         Bill bill = Bill.creator()
-                .setIdentificationField("25794150099003551916515000211407100000000000000")
-                .create();
+            .setIdentificationField("25794150099003551916515000211407100000000000000")
+            .create();
     }
 
     private static void pixStaticQrCode() {
         PixQrCode qrCode = PixQrCode
-                .creator()
-                .setAddressKey(PixAddressKey.reader().read().getData().get(0).key)
-                .setDescription("teste")
-                .setValue(Money.create(0.01))
-                .create();
+            .creator()
+            .setAddressKey(PixAddressKey.reader().read().getData().get(0).key)
+            .setDescription("teste")
+            .setValue(Money.create(0.01))
+            .create();
         System.out.printf(qrCode.payload);
     }
 
     private static void payment() {
         Payment payment = Payment.creator()
-                .setCustomer("cus_000072683114")
-                .setBillingType(BillingType.PIX)
-                .setDueDate(new Date())
-                .setInstallmentCount(2)
-                .setInstallmentValue(Money.create(50))
-                .setDescription("Teste")
-                .create();
+            .setCustomer("cus_000072683114")
+            .setBillingType(BillingType.PIX)
+            .setDueDate(new Date())
+            .setInstallmentCount(2)
+            .setInstallmentValue(Money.create(50))
+            .setDescription("Teste")
+            .create();
 
         ResourceSet<Payment> paymentResourceSet = Payment.reader()
-                .setStatus(PaymentStatus.RECEIVED)
-                .setStartPaymentDate(new Date())
-                .setFinishDueDate(new Date())
-                .read();
+            .setStatus(PaymentStatus.RECEIVED)
+            .setStartPaymentDate(new Date())
+            .setFinishDueDate(new Date())
+            .read();
 
         DeletedResource paymentDeleted = Payment.deleter(payment.getId()).delete();
         payment = Payment.restorer(payment.getId()).create();
@@ -254,9 +203,9 @@ public class Examples {
     private static void customerAccount() {
         CustomerAccount.fetcher("cus_000072683044").fetch();
         CustomerAccount customerAccount = CustomerAccount.creator()
-                .setName("criado via API")
-                .setCpfCnpj("10030823005")
-                .create();
+            .setName("criado via API")
+            .setCpfCnpj("10030823005")
+            .create();
     }
 
     private static void notification() {
@@ -267,57 +216,57 @@ public class Examples {
     private static void paymentLink() {
         PaymentLink link = PaymentLink.fetcher("725104409743").fetch();
         ResourceSet<PaymentLink> paymentLinkResourceSet = PaymentLink
-                .reader()
-                .read();
+            .reader()
+            .read();
         DeletedResource deletedPaymentLink = PaymentLink
-                .deleter("725104409743")
-                .delete();
+            .deleter("725104409743")
+            .delete();
 
         PaymentLink paymentLink = PaymentLink.creator()
-                .setName("name")
-                .setBillingType(BillingType.PIX)
-                .setChargeType(PaymentLinkChargeType.INSTALLMENT)
-                .setEndDate(new Date())
-                .setDueDateLimitDays(10)
-                .setMaxInstallmentCount(2)
-                .create();
+            .setName("name")
+            .setBillingType(BillingType.PIX)
+            .setChargeType(PaymentLinkChargeType.INSTALLMENT)
+            .setEndDate(new Date())
+            .setDueDateLimitDays(10)
+            .setMaxInstallmentCount(2)
+            .create();
 
         PaymentLink updated = PaymentLink.updater(paymentLink.getId())
-                .setName("name")
-                .setBillingType(BillingType.PIX)
-                .setChargeType(PaymentLinkChargeType.INSTALLMENT)
-                .setEndDate(new Date())
-                .setDueDateLimitDays(10)
-                .setMaxInstallmentCount(2)
-                .update();
+            .setName("name")
+            .setBillingType(BillingType.PIX)
+            .setChargeType(PaymentLinkChargeType.INSTALLMENT)
+            .setEndDate(new Date())
+            .setDueDateLimitDays(10)
+            .setMaxInstallmentCount(2)
+            .update();
     }
 
     private static void financialTransaction() {
         ResourceSet<FinancialTransaction> financialTransactionResourceSet = FinancialTransaction
-                .reader()
-                .setTransferId("transferId")
-                .read();
+            .reader()
+            .setTransferId("transferId")
+            .read();
     }
 
     private static void invoice() {
         Invoice invoice = Invoice.creator()
-                .setServiceDescription("Nota fiscal da Fatura 101940. Descrição dos Serviços: ANÁLISE E DESENVOLVIMENTO DE SISTEMAS")
-                .setObservations("Mensal referente aos trabalhos de Junho.")
-                .setValue(Money.create(300))
-                .setDeductions(Money.create(2))
-                .setEffectiveDate(new Date())
-                .setMunicipalServiceName("Análise e Desenvolvimento de Sistemas")
-                .setTaxes(
-                        new Taxes()
-                                .setRetainIss(true)
-                                .setIss(Money.create(3))
-                                .setCofins(Money.create(3))
-                                .setCsll(Money.create(1))
-                                .setInss(Money.create(3))
-                                .setIr(Money.create(1.5))
-                                .setPis(Money.create(0.65))
-                )
-                .create();
+            .setServiceDescription("Nota fiscal da Fatura 101940. Descrição dos Serviços: ANÁLISE E DESENVOLVIMENTO DE SISTEMAS")
+            .setObservations("Mensal referente aos trabalhos de Junho.")
+            .setValue(Money.create(300))
+            .setDeductions(Money.create(2))
+            .setEffectiveDate(new Date())
+            .setMunicipalServiceName("Análise e Desenvolvimento de Sistemas")
+            .setTaxes(
+                new Taxes()
+                    .setRetainIss(true)
+                    .setIss(Money.create(3))
+                    .setCofins(Money.create(3))
+                    .setCsll(Money.create(1))
+                    .setInss(Money.create(3))
+                    .setIr(Money.create(1.5))
+                    .setPis(Money.create(0.65))
+            )
+            .create();
     }
 
     private static void finance() {
@@ -332,8 +281,8 @@ public class Examples {
     private static void commercialInfo() {
         CommercialInfo commercialInfo = CommercialInfo.fetcher().fetch();
         CommercialInfo updatedCommercialInfo = CommercialInfo.updater()
-                .setSite("https://yourSite.com.br")
-                .update();
+            .setSite("https://yourSite.com.br")
+            .update();
     }
 
     private static void accountNumber() {
@@ -351,13 +300,13 @@ public class Examples {
 
     private static void subAccount() {
         Account account = Account.creator()
-                .setName("Teste sub conta")
-                .setBirthDate(new Date())
-                .setCompanyType("LIMITED")
-                .setEmail("joaoexample2@gmail.com")
-                .setPostalCode("36572122")
-                .setCpfCnpj("87.326.705/0001-81")
-                .create();
+            .setName("Teste sub conta")
+            .setBirthDate(new Date())
+            .setCompanyType("LIMITED")
+            .setEmail("joaoexample2@gmail.com")
+            .setPostalCode("36572122")
+            .setCpfCnpj("87.326.705/0001-81")
+            .create();
         System.out.println(account.getName());
 
         Account fetched = Account.fetcher(account.id).fetch();
@@ -366,26 +315,26 @@ public class Examples {
 
     private static void subscription() {
         Subscription subscription = Subscription.creator()
-                .setCustomer("cus_000072683044")
-                .setCycle(SubscriptionCycle.MONTHLY)
-                .setNextDueDate(new Date())
-                .setValue(Money.create(100))
-                .setDescription("Teste")
-                .create();
+            .setCustomer("cus_000072683044")
+            .setCycle(SubscriptionCycle.MONTHLY)
+            .setNextDueDate(new Date())
+            .setValue(Money.create(100))
+            .setDescription("Teste")
+            .create();
     }
 
     private static Account createFirstAccountIfNecessary() {
         List<Account> accounts = Account.reader().read().getData();
         if (accounts.isEmpty()) {
             return Account.creator()
-                    .setName("Teste sub conta")
-                    .setBirthDate(new Date())
-                    .setCompanyType("LIMITED")
-                    .setEmail("joaoexample2@gmail.com")
-                    .setPostalCode("36572122")
-                    .setIncomeValue(Money.create(1000))
-                    .setCpfCnpj("87.326.705/0001-81")
-                    .create();
+                .setName("Teste sub conta")
+                .setBirthDate(new Date())
+                .setCompanyType("LIMITED")
+                .setEmail("joaoexample2@gmail.com")
+                .setPostalCode("36572122")
+                .setIncomeValue(Money.create(1000))
+                .setCpfCnpj("87.326.705/0001-81")
+                .create();
         }
         return accounts.get(0);
     }
